refactor: extract shared menu overlay logic to the mixin
