+++ packages/context-menu/theme/material/vaadin-context-menu-overlay-styles.js
@@ -11,3 +11,5 @@ const contextMenuOverlay = css`
 registerStyles('vaadin-context-menu-overlay', [menuOverlay, contextMenuOverlay], {
   moduleId: 'material-context-menu-overlay',
 });
+
+export { contextMenuOverlay };
