+++ packages/context-menu/theme/material/vaadin-context-menu.js
@@ -1,5 +1,5 @@
 import './vaadin-context-menu-item-styles.js';
 import './vaadin-context-menu-list-box-styles.js';
-import './vaadin-context-menu-styles.js';
+import './vaadin-context-menu-overlay-styles.js';
 import '@vaadin/overlay/theme/material/vaadin-overlay.js';
 import '../../src/vaadin-context-menu.js';
