+++ packages/menu-bar/src/vaadin-menu-bar-overlay.js
@@ -3,15 +3,21 @@
  * Copyright (c) 2019 - 2023 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { ContextMenuOverlay } from '@vaadin/context-menu/src/vaadin-context-menu-overlay.js';
+import { MenuOverlayMixin } from '@vaadin/context-menu/src/vaadin-menu-overlay-mixin.js';
+import { styles } from '@vaadin/context-menu/src/vaadin-menu-overlay-styles.js';
+import { Overlay } from '@vaadin/overlay/src/vaadin-overlay.js';
+import { registerStyles } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+
+registerStyles('vaadin-menu-bar-overlay', styles, { moduleId: 'vaadin-menu-bar-overlay-styles' });
 
 /**
  * An element used internally by `<vaadin-menu-bar>`. Not intended to be used separately.
  *
- * @extends ContextMenuOverlay
+ * @extends Overlay
+ * @mixes MenuOverlayMixin
  * @private
  */
-class MenuBarOverlay extends ContextMenuOverlay {
+class MenuBarOverlay extends MenuOverlayMixin(Overlay) {
   static get is() {
     return 'vaadin-menu-bar-overlay';
   }
