+++ packages/menu-bar/theme/lumo/vaadin-menu-bar-overlay-styles.js
@@ -1,11 +1,13 @@
+import { contextMenuOverlay } from '@vaadin/context-menu/theme/lumo/vaadin-context-menu-overlay-styles.js';
+import { menuOverlay } from '@vaadin/vaadin-lumo-styles/mixins/menu-overlay.js';
 import { css, registerStyles } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
 
-registerStyles(
-  'vaadin-menu-bar-overlay',
-  css`
-    :host(:first-of-type) {
-      padding-top: var(--lumo-space-xs);
-    }
-  `,
-  { moduleId: 'lumo-menu-bar-overlay' },
-);
+const menuBarOverlay = css`
+  :host(:first-of-type) {
+    padding-top: var(--lumo-space-xs);
+  }
+`;
+
+registerStyles('vaadin-menu-bar-overlay', [menuOverlay, contextMenuOverlay, menuBarOverlay], {
+  moduleId: 'lumo-menu-bar-overlay',
+});
