+++ packages/menu-bar/theme/lumo/vaadin-menu-bar.js
@@ -3,5 +3,5 @@ import './vaadin-menu-bar-item-styles.js';
 import './vaadin-menu-bar-list-box-styles.js';
 import './vaadin-menu-bar-overlay-styles.js';
 import './vaadin-menu-bar-styles.js';
-import '@vaadin/context-menu/theme/lumo/vaadin-context-menu.js';
+import '@vaadin/overlay/theme/lumo/vaadin-overlay.js';
 import '../../src/vaadin-menu-bar.js';
